Toggle the chat work panel from the header
A pull-request button sits left of the chat settings menu whenever the
agent has posted a PR, toggling the top-right work module (tinted
primary while open; each chat starts with it visible). The panel's
card now uses the standard rich PR card style, matching PR link cards
elsewhere.

Splits ChatList out of ChatsScreen (file size ceiling) with the
shared-chat metadata check moved to lib/chatShared.
